Fixed UseEffect hook that contains async call for production build
When deploying to Vercel or creating production build locally, the produced html had almost no components inside <main>, in other words, the html was almost empty. This was caused by the UseEffect hook that was calling the async functions that were inside it. Fixed by removing the return statement inside the UseEffect hook.